refactor: update config import paths in test files
