+++ tests/admin.test.ts
@@ -5,7 +5,7 @@ import {
   type AdminLeaderboardEntry,
 } from '../src/utils/admin'
 import type { LeaderboardEntry } from '../src/utils/leaderboard'
-import { WORKER_BASE_URL } from '../src/utils/leaderboardConfig'
+import { WORKER_BASE_URL } from '../src/utils/config'
 import { storage, fetchCalls, mockFetch, resetMocks } from './helpers'
 
 beforeEach(() => resetMocks())
